% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgrecap.R
\docType{package}
\name{hgrecap-package}
\alias{hgrecap-package}
\title{Bayesian Changepoint for Binomial Data}
\description{
This package provides a binomial variant of bayesian changepoint analysis.
}
\author{
Neal Fultz \email{njf@zestfinance.com},
}

