% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgrecap.R
\name{hgrecap.control}
\alias{hgrecap.control}
\title{MCMC control parameters}
\usage{
hgrecap.control(mcmc.iterations = 500, mcmc.thin = 5, mcmc.burnin = 50,
  verbose = 0)
}
\arguments{
\item{mcmc.iterations}{Number of iterations to run}

\item{mcmc.thin}{an integer > 0, Number of iterations to thin between samples}

\item{verbose}{- Not used right now.}

\item{mcmd.burnin}{an integer >= 0}
}
\description{
MCMC control parameters
}

