% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgrecap.R
\name{hgrecap}
\alias{hgrecap}
\title{Hypergeometric Recapture Sampler}
\usage{
hgrecap(s, prior = list(lambda = 0), mcmc.control = hgrecap.control())
}
\arguments{
\item{s}{a list of observation ids}

\item{prior}{a list specifying the lambda}

\item{mcmc.control}{see \code{\link{hgrecap.control}}}
}
\value{
a list containing samples and transition probabilities
}
\description{
Uses MCMC to sample the size.
}

